requirements.txt
firebase-admin==6.2.0
flask==2.3.3
python-dotenv==1.0.0
flask-cors==4.0.0
pydantic==2.4.2
python-jose==3.3.0
bcrypt==4.0.1 